feat(dashboard): add client dashboard and minhas-solicitacoes pages with dynamic imports
- Added DashboardClient component with user role based content and approval actions
- Refactored dashboard page to dynamically import DashboardClient
- Added MinhasSolicitacoesClient component to fetch and display user solicitations
- Refactored minhas-solicitacoes page to dynamically import MinhasSolicitacoesClient
- Added ClientOnly component for client-side rendering control
- Added force-dynamic export to some pages for dynamic rendering
- Minor imports and code cleanup in calcular-rotas and admin usuarios pages
