+++ src/app/admin/usuarios/page.tsx
@@ -1,5 +1,7 @@
 'use client'
 
+export const dynamic = 'force-dynamic'
+
 import { useState, useEffect } from 'react'
 import { useAuth } from '@/hooks/useAuth'
 import { useUsuarios, useClientes } from '@/hooks/useApi'
