+++ src/components/ClientOnly.tsx
@@ -1,0 +1,22 @@
+'use client'
+
+import { useEffect, useState } from 'react'
+
+interface ClientOnlyProps {
+  children: React.ReactNode
+  fallback?: React.ReactNode
+}
+
+export default function ClientOnly({ children, fallback = <div>Loading...</div> }: ClientOnlyProps) {
+  const [hasMounted, setHasMounted] = useState(false)
+
+  useEffect(() => {
+    setHasMounted(true)
+  }, [])
+
+  if (!hasMounted) {
+    return <>{fallback}</>
+  }
+
+  return <>{children}</>
+}
